Juntar objetos de estados con objetos de player
