+++ EstadosManager.cpp
@@ -1,0 +1,55 @@
+#include "EstadosManager.hpp"
+
+namespace Crazy
+{
+    void EstadosManager::Anyadir( Estado* nuevoEstado, bool reemplazar)
+    {
+        anyadiendo = true;
+        reemplazando = reemplazar;
+        estado = move(nuevoEstado);
+    }
+    
+    void EstadosManager::Eliminar()
+    {
+        eliminando = true;
+    }
+    
+    void EstadosManager::ProcesarPilaEstados()
+    {
+        if ( eliminando && !estados.empty())
+        {
+            estados.pop();
+            
+            if (!estados.empty())
+            {
+                estados.top()->Reanudar();
+            }
+            eliminando = false;
+        }
+        
+        if (anyadiendo)
+        {
+            if (!estados.empty())
+            {
+                if (reemplazando)
+                {
+                    estados.pop();
+                }
+                else
+                {
+                    estados.top()->Pausar();
+                }
+            }
+            
+            estados.push(move(estado));
+            estado=0;
+            estados.top()->Init();
+            anyadiendo = false;
+        }
+    }
+    
+    Estado* &EstadosManager::GetEstadoActivo()
+    {
+        return estados.top();
+    }
+}
